v1.0.0

+++ scVAEIT/VAEIT.py
@@ -20,34 +20,34 @@ import scanpy as sc
 
 class VAEIT():
     """
-    Variational Inference for Trajectory by AutoEncoder.
+    Variational Inference for integration and transfer learning.
     """
-    def __init__(self, config: SimpleNamespace, data, masks, id_dataset=None, batches_cate=None, batches_cont=None
+    def __init__(self, config: SimpleNamespace, data, masks, id_dataset=None, batches_cate=None, batches_cont=None, conditions=None
         ):
         '''
-        Get input data for model.
-
+        Initialize the VAEIT model with the provided configuration and data.
 
         Parameters
         ----------
         config : SimpleNamespace
-            Dict of config.
+            Configuration object containing model parameters.
         data : np.array
-            The cell-by-feature matrix.
+            The cell-by-feature matrix representing the input data.
         masks : np.array
-            Masks that indicate missingness. 1 is missing and 0 is observed.
-            It can be the full mask matrix with the same shape as `data`, or a condensed matrix that can be indexed 
-            using values in `id_dataset`.
+            Masks indicating missingness, where 1 represents missing and 0 represents observed.
+            This can be a full mask matrix with the same shape as `data`, or a condensed matrix indexed by `id_dataset`.
         id_dataset : np.array, optional
-            The dataset integer id for each cell. If masks is a condensed matrix, it is required.
+            Integer IDs for each cell in the dataset. Required if `masks` is a condensed matrix.
         batches_cate : np.array, optional
-            The categorical batch information for each cell.
+            Categorical batch information for each cell.
         batches_cont : np.array, optional
-            The continuous batch information for each cell.
+            Continuous batch information for each cell.
+        conditions : np.array, optional
+            Conditions for each cell, used for calculating MMD loss if `config.gamma` is not 0.
 
         Returns
         -------
-        None.
+        None
         '''
         self.dict_method_scname = {
             'PCA' : 'X_pca',
@@ -60,16 +60,20 @@ class VAEIT():
         self.data = tf.convert_to_tensor(data, dtype=tf.keras.backend.floatx())
         
         # prprocessing config
+        if 'dim_input_arr' not in config:
+            config['dim_input_arr'] = data.shape[1]
         if np.isscalar(config['dim_input_arr']):
             config['dim_input_arr'] = np.array([config['dim_input_arr']], dtype=np.int32)
         n_modal = len(config['dim_input_arr'])
-        n_block = len(config['dim_block'])
+        n_block = len(config['dim_block']) if 'dim_block' in config else n_modal
+        
+        
         config = dict(filter(lambda item: item[1] is not None, config.items()))
 
 
         config = {**{
             'beta_kl':2., # weight for beta-VAE
-            'beta_reverse':0.2, # weight for reverse prediction (use masked out features to predict the observed features)
+            'beta_reverse':0., # weight for reverse prediction (use masked out features to predict the observed features)
             'beta_modal':np.ones(n_modal, dtype=np.float32), # weight for each modality
             'p_modal':None,
 
@@ -82,7 +86,9 @@ class VAEIT():
 
 
             'skip_conn':False, # whether to use skip connection in decoder                     
-            'max_vals':tf.constant(tf.reduce_max(tf.math.abs(self.data)), shape=self.data.shape[1], dtype=tf.keras.backend.floatx())
+            'max_vals':tf.constant(tf.reduce_max(tf.math.abs(self.data)), dtype=tf.keras.backend.floatx()),
+            
+            'gamma':0.
         }, **config}
 
         if isinstance(config, dict):
@@ -96,12 +102,6 @@ class VAEIT():
             config.dim_block_embed = np.full(n_block, config.dim_block_embed, dtype=np.int32)
         config.dim_block_enc = check_arr_type(config.dim_block_enc, np.int32)
         config.dim_block_dec = check_arr_type(config.dim_block_dec, np.int32)
-
-        if np.isscalar(config.max_vals):
-            config.max_vals = tf.constant(config.max_vals, shape=self.data.shape[1], dtype=tf.keras.backend.floatx())
-        else:
-            config.max_vals = tf.convert_to_tensor(config.max_vals, dtype=tf.keras.backend.floatx())
-            # config.max_vals = tf.reduce_max(tf.math.abs(self.data), axis=0)
         
         self.config = config
 
@@ -109,7 +109,7 @@ class VAEIT():
         self.batches = np.array([], dtype=np.float32).reshape((data.shape[0],0))
         if batches_cate is not None:
             batches_cate = np.array(batches_cate)
-            self.cat_enc = OneHotEncoder().fit(batches_cate)
+            self.cat_enc = OneHotEncoder(drop='first').fit(batches_cate)
             self.batches = self.cat_enc.transform(batches_cate).toarray()
         if batches_cont is not None:
             self.cont_enc = StandardScaler()            
@@ -118,7 +118,19 @@ class VAEIT():
             batches_cont = np.array(batches_cont, dtype=np.float32)            
             self.batches = np.c_[self.batches, batches_cont]
         self.batches = tf.convert_to_tensor(self.batches, dtype=tf.keras.backend.floatx())
+        
+        if conditions is not None and self.config.gamma != 0.:
+            ## observations with np.nan will not participant in calculating mmd_loss
+            conditions = np.array(conditions)
+            if len(conditions.shape)<2:
+                conditions = conditions[:,None]
+            self.conditions = OrdinalEncoder(dtype=int, encoded_missing_value=-1).fit_transform(conditions) + int(1)
             
+        else:
+            self.config.gamma = 0.
+            self.conditions = np.zeros((data.shape[0],1), dtype=np.int32)
+        self.conditions = tf.cast(self.conditions, tf.int32)
+
         # [num_cells, num_features]
         self.masks = tf.convert_to_tensor(masks, dtype=tf.keras.backend.floatx())
                 
@@ -129,8 +141,10 @@ class VAEIT():
             self.id_dataset = tf.convert_to_tensor(id_dataset, dtype=tf.int32)
             self.full_masks = False
         
+        self.batch_size_inference = 512
+        
         self.reset()
-        print(self.config, self.masks.shape, self.data.shape, self.batches.shape)
+        print(self.config, self.masks.shape, self.data.shape, self.batches.shape, flush = True)
         
 
     def reset(self):
@@ -144,50 +158,65 @@ class VAEIT():
 
     def train(self, valid = False, stratify = False, test_size = 0.1, random_state: int = 0,
             learning_rate: float = 1e-3, batch_size: Optional[int] = None, batch_size_inference: Optional[int] = None, 
-            L: int = 1, num_epoch: int = 200, num_step_per_epoch: Optional[int] = None, save_every_epoch: Optional[int] = 25,
+            L: int = 1, num_epoch: int = 200, num_step_per_epoch: Optional[int] = None, save_every_epoch: Optional[int] = 25, init_epoch: Optional[int] = 1,
             early_stopping_patience: int = 10, early_stopping_tolerance: float = 1e-4, 
             early_stopping_relative: bool = True, verbose: bool = False,
-            checkpoint_dir: Optional[str] = None, delete_existing: Optional[str] = True, eval_func=None): 
-        '''Pretrain the model with specified learning rate.
+            checkpoint_dir: Optional[str] = None, delete_existing: Optional[str] = True, eval_func = None): 
+        '''
+        Train the VAEIT model with the specified parameters.
 
         Parameters
         ----------
+        valid : bool, optional
+            Whether to use a validation set during training. Default is False.
+        stratify : bool, optional
+            Whether to stratify the split when creating the validation set. Default is False.
         test_size : float or int, optional
-            The proportion or size of the test set.
+            The proportion or size of the validation set. Default is 0.1.
         random_state : int, optional
-            The random state for data splitting.
+            The random state for data splitting. Default is 0.
         learning_rate : float, optional
-            The initial learning rate for the Adam optimizer.
+            The initial learning rate for the Adam optimizer. Default is 1e-3.
         batch_size : int, optional 
             The batch size for training. Default is 256 when using full mask matrices, or 64 otherwise.
         batch_size_inference : int, optional
             The batch size for inference. Default is 256 when using full mask matrices, or 64 otherwise.
         L : int, optional 
-            The number of MC samples.
+            The number of Monte Carlo samples. Default is 1.
         num_epoch : int, optional 
-            The maximum number of epochs.
+            The maximum number of epochs. Default is 200.
         num_step_per_epoch : int, optional 
-            The number of step per epoch, it will be inferred from number of cells and batch size if it is None.            
+            The number of steps per epoch. If None, it will be inferred from the number of cells and batch size. Default is None.
+        save_every_epoch : int, optional 
+            Frequency (in epochs) to save model checkpoints. Default is 25.
+        init_epoch : int, optional 
+            The initial epoch number. Default is 1.
         early_stopping_patience : int, optional 
-            The maximum number of epochs if there is no improvement.
+            The number of epochs to wait for improvement before early stopping. Default is 10.
         early_stopping_tolerance : float, optional 
-            The minimum change of loss to be considered as an improvement.
+            The minimum change in loss to be considered as an improvement. Default is 1e-4.
         early_stopping_relative : bool, optional
-            Whether monitor the relative change of loss as stopping criteria or not.
-        path_to_weights : str, optional 
-            The path of weight file to be saved; not saving weight if None.
+            Whether to monitor the relative change in loss for early stopping. Default is True.
+        verbose : bool, optional
+            Whether to print the training process. Default is False.
+        checkpoint_dir : str, optional 
+            Directory to save model checkpoints. Default is None.
+        delete_existing : bool, optional 
+            Whether to delete existing checkpoints in the directory. Default is True.
+        eval_func : function, optional
+            A function to evaluate the model, which takes the VAE as an input. Default is None.
 
         Returns
         -------
         hist : dict
-            The history of loss.
+            A dictionary containing the history of training and validation losses.
         '''
 
         if batch_size is None:
             batch_size = 256 if self.full_masks else 64
         if batch_size_inference is None:
             batch_size_inference = batch_size
-
+            
         if valid:
             if stratify is False:
                 stratify = None    
@@ -199,26 +228,26 @@ class VAEIT():
                                     random_state=random_state)
             
             self.dataset_train = tf.data.Dataset.from_tensor_slices((
-                self.data[id_train], self.batches[id_train], self.id_dataset[id_train]
+                self.data[id_train], self.batches[id_train], self.id_dataset[id_train], self.conditions[id_train]
                 )).shuffle(buffer_size = len(id_train), seed=0,
-                           reshuffle_each_iteration=True).batch(batch_size).prefetch(tf.data.experimental.AUTOTUNE)
+                           reshuffle_each_iteration=True).batch(batch_size, drop_remainder=True).prefetch(tf.data.experimental.AUTOTUNE)
 
             self.dataset_valid = tf.data.Dataset.from_tensor_slices((
-                    self.data[id_valid], self.batches[id_valid], self.id_dataset[id_valid]
+                    self.data[id_valid], self.batches[id_valid], self.id_dataset[id_valid], self.conditions[id_valid]
                 )).batch(batch_size_inference).prefetch(tf.data.experimental.AUTOTUNE)
         else:
             id_train = np.arange(self.data.shape[0])
             self.dataset_train = tf.data.Dataset.from_tensor_slices((
-                self.data, self.batches, self.id_dataset
+                self.data, self.batches, self.id_dataset, self.conditions
                 )).shuffle(buffer_size = len(id_train), seed=0,
-                           reshuffle_each_iteration=True).batch(batch_size).prefetch(tf.data.experimental.AUTOTUNE)
+                           reshuffle_each_iteration=True).batch(batch_size, drop_remainder=True).prefetch(tf.data.experimental.AUTOTUNE)
             self.dataset_valid = None
             
         if num_step_per_epoch is None:
-            num_step_per_epoch = len(id_train)//batch_size+1
+            num_step_per_epoch = len(id_train)//batch_size
             
         if checkpoint_dir is not None:        
-            if delete_existing and tf.io.gfile.exists(checkpoint_dir):
+            if init_epoch==1 and delete_existing and tf.io.gfile.exists(checkpoint_dir):
                 print("Deleting old log directory at {}".format(checkpoint_dir))
                 tf.io.gfile.rmtree(checkpoint_dir)
             tf.io.gfile.makedirs(checkpoint_dir)
@@ -229,10 +258,11 @@ class VAEIT():
             self.vae,
             checkpoint_dir,
             learning_rate,                        
-            L,
-            num_epoch,
-            num_step_per_epoch,
-            save_every_epoch,
+            int(L),
+            int(num_epoch),
+            int(num_step_per_epoch),
+            int(save_every_epoch),
+            init_epoch,
             early_stopping_patience,
             early_stopping_tolerance,
             early_stopping_relative,
@@ -249,16 +279,28 @@ class VAEIT():
             checkpoint, path_to_weights, max_to_keep=None
         )
         save_path = manager.save()        
-        print("Saved checkpoint: {}".format(save_path))
+        print("Saved checkpoint: {}".format(save_path), flush = True)
         
 
     def load_model(self, path_to_weights):
         checkpoint = tf.train.Checkpoint(net=self.vae)
         status = checkpoint.restore(path_to_weights)
-        print("Loaded checkpoint: {}".format(status))
+        print("Loaded checkpoint: {}".format(status), flush = True)
+
+
+    def set_dataset(self, batch_size_inference=512):
+        if not hasattr(self, 'dataset_full') or batch_size_inference != self.batch_size_inference:
+            id_samples = np.arange(self.data.shape[0])
+            self.dataset_full = tf.data.Dataset.from_tensor_slices((
+                    self.data, self.batches, self.id_dataset, tf.constant(id_samples)
+                )).shuffle(
+                    buffer_size = len(id_samples), seed=0,
+                    reshuffle_each_iteration=True
+                ).batch(batch_size_inference).prefetch(tf.data.experimental.AUTOTUNE)
+            self.batch_size_inference = batch_size_inference
 
             
-    def get_latent_z(self, masks=None, batch_size_inference=512):
+    def get_latent_z(self, masks=None, zero_out=True, batch_size_inference=512):
         ''' get the posterier mean of current latent space z (encoder output)
 
         Returns
@@ -266,27 +308,20 @@ class VAEIT():
         z : np.array
             \([N,d]\) The latent means.
         ''' 
-        if not hasattr(self, 'dataset_full'):
-            self.dataset_full = tf.data.Dataset.from_tensor_slices((
-                    self.data, self.batches, self.id_dataset
-                )).batch(batch_size_inference).prefetch(tf.data.experimental.AUTOTUNE)
-
-        return self.vae.get_z(self.dataset_full, self.full_masks, masks)
+        self.set_dataset(batch_size_inference)
+        return self.vae.get_z(self.dataset_full, self.full_masks, masks, zero_out)
 
 
     def get_denoised_data(self, masks=None, zero_out=True, batch_size_inference=512, return_mean=True, L=50):
-        if not hasattr(self, 'dataset_full'):
-            self.dataset_full = tf.data.Dataset.from_tensor_slices((
-                    self.data, self.batches, self.id_dataset
-                )).batch(batch_size_inference).prefetch(tf.data.experimental.AUTOTUNE)
+        self.set_dataset(batch_size_inference)        
 
         return self.vae.get_recon(self.dataset_full, self.full_masks, masks, zero_out, return_mean, L)
     
         
-    def update_z(self, masks=None, batch_size_inference=512):
-        self.z = self.get_latent_z(masks, batch_size_inference)
+    def update_z(self, masks=None, zero_out=True, batch_size_inference=512, **kwargs):
+        self.z = self.get_latent_z(masks, zero_out, batch_size_inference)
         self.adata = sc.AnnData(self.z)
-        sc.pp.neighbors(self.adata)
+        sc.pp.neighbors(self.adata, **kwargs)
 
     
     def visualize_latent(self, method: str = "UMAP", 
@@ -343,14 +378,3 @@ class VAEIT():
             axes = sc.pl.draw_graph(self.adata, color = color, **kwargs)
             
         return axes
-
-
-
-
-
-
-
-
- 
-
-
